+++ Functions/Responses/BaseResponse.cs
@@ -9,9 +9,15 @@ namespace Supabase.Functions.Responses
     /// </summary>
     public class BaseResponse
     {
+        /// <summary>
+        /// The response message
+        /// </summary>
         [JsonIgnore]
         public HttpResponseMessage? ResponseMessage { get; set; }
 
+        /// <summary>
+        /// The response content.
+        /// </summary>
         [JsonIgnore]
         public string? Content { get; set; }
     }